Add feature to import pages from local directory hierarchy
This adds a maintenance script that scans a directory for files
and imports them. Top-level files are imported into the main
namespace, and files in subdirectories are imported into namespaces
named as those subdirectories.

No file extensions are stripped (in order to be able to import JS
and CSS files etc.).

A 'watch' option is added, which will keep the script running and
monitoring the files for changes. Whenever one is modified it will
be re-imported. This feature relies on the inotify Pecl extension.

Also fix a couple of coding-standard errors elsewhere.

Change-Id: I872dddfe43eaf6029d5282ea18e9fca426fd8995